Update mi600-webif.php
After some time of operation, an error has appeared that always occurs when 123solar is restarted during the period when the inverter is offline. In this case there was a short spike in the power chart because the KWHT start value of the inverter was not correctly preset with the last value. Thus in the inverter datafiles the KWHT value was entered with 0 instead of the correct value currently stored in the inverter. This fix eliminates this behavior.